backend/requirements.txt
requests
beautifulsoup4
pandas
supabase
yt-dlp
python-dotenv
aiohttp
schedule
google-generativeai
fuzzywuzzy
python-levenshtein
